# apps/express-api/public/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Macro AI Express API",
		"version": "0.0.1",
		"description": "API documentation for Macro AI\n\t\t\t\t\n## Rate Limiting\nThis API implements rate limiting to protect against abuse:\n\n- **Global Rate Limit**: 100 requests per 15 minutes for all endpoints\n- **Authentication Rate Limit**: 10 requests per hour for authentication endpoints\n- **API Rate Limit**: 60 requests per minute for API endpoints\n\nRate limit headers are included in responses to help track usage.",
		"license": {
			"name": "MIT",
			"url": "https://spdx.org/licenses/MIT.html"
		}
	},
	"servers": [
		{
			"url": "http://localhost:3030/api"
		}
	],
	"components": {
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "macro-ai-accessToken"
			},
			"apiKey": {
				"type": "apiKey",
				"in": "header",
				"name": "x-api-key"
			}
		},
...
	},
	"paths": {
		"/auth/register": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Register new user",
				"description": "Creates a new user account with email and password. Sends confirmation email to verify the account.",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email",
										"description": "User email address"
									},
									"password": {
										"type": "string",
										"minLength": 8,
										"maxLength": 15,
										"pattern": "\\d",
										"description": "User password"
									},
									"confirmPassword": {
										"type": "string",
										"minLength": 8,
										"maxLength": 15,
										"pattern": "\\d",
										"description": "Confirm password"
									}
								},
								"required": ["email", "password", "confirmPassword"]
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "User registered successfully - confirmation email sent",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Response message"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"400": {
						"description": "Invalid request data - validation failed or passwords do not match",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Validation error message",
											"example": "Validation Failed"
										},
										"details": {
											"type": "object",
											"additionalProperties": {
												"nullable": true
											},
											"description": "Detailed validation error information",
											"example": {
												"field": "email",
												"message": "Invalid email format"
											}
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"409": {
						"description": "Conflict - User already exists (from database check)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Conflict error message",
											"example": "Resource conflict"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"422": {
						"description": "User already exists in Cognito (UsernameExistsException)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "User already exists error message",
											"example": "User already exists"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"429": {
						"description": "Too many registration attempts - rate limit exceeded",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "number",
											"description": "HTTP status code",
											"example": 429
										},
										"message": {
											"type": "string",
											"description": "Rate limit error message",
											"example": "Too many requests, please try again later."
										}
									},
									"required": ["status", "message"]
								}
							}
						}
					},
					"500": {
						"description": "Server error - Cognito or database error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Internal server error message",
											"example": "Internal server error"
										}
									},
									"required": ["message"]
								}
							}
						}
					}
				}
			}
		},
		"/auth/confirm-registration": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Confirm user registration",
				"description": "Confirms user registration using the confirmation code sent via email.",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email",
										"description": "User email address"
									},
									"code": {
										"type": "number",
										"description": "Verification code sent to email"
									}
								},
								"required": ["email", "code"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "User registration confirmed successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Response message"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"400": {
						"description": "Invalid confirmation code (CodeMismatchException) or request data validation failed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Invalid verification code error message",
											"example": "Invalid verification code"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"403": {
						"description": "Forbidden - User already confirmed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Forbidden error message",
											"example": "Forbidden"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Not found error message",
											"example": "Resource not found"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"410": {
						"description": "Verification code has expired (ExpiredCodeException)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Expired verification code error message",
											"example": "Verification code has expired"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"429": {
						"description": "Too many confirmation attempts - rate limit exceeded",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "number",
											"description": "HTTP status code",
											"example": 429
										},
										"message": {
											"type": "string",
											"description": "Rate limit error message",
											"example": "Too many requests, please try again later."
										}
									},
									"required": ["status", "message"]
								}
							}
						}
					},
					"500": {
						"description": "Server error - Cognito or database error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Internal server error message",
											"example": "Internal server error"
										}
									},
									"required": ["message"]
								}
							}
						}
					}
				}
			}
		},
		"/auth/resend-confirmation-code": {
			"post": {
				"tags": ["Authentication"],
				"summary": "Resend confirmation code",
				"description": "Resends the confirmation code to the user's email address for account verification.",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email",
										"description": "User email address"
									}
								},
								"required": ["email"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Confirmation code resent successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Response message"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"400": {
						"description": "Invalid request data - validation failed or invalid email format",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Validation error message",
											"example": "Validation Failed"
										},
										"details": {
											"type": "object",
											"additionalProperties": {
												"nullable": true
											},
											"description": "Detailed validation error information",
											"example": {
												"field": "email",
												"message": "Invalid email format"
											}
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"403": {
						"description": "Forbidden - User already confirmed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Forbidden error message",
											"example": "Forbidden"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Not found error message",
											"example": "Resource not found"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"429": {
						"description": "Too many resend attempts - rate limit exceeded",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "number",
											"description": "HTTP status code",
											"example": 429
										},
										"message": {
											"type": "string",
											"description": "Rate limit error message",
											"example": "Too many requests, please try again later."
										}
									},
									"required": ["status", "message"]
								}
							}
						}
					},
					"500": {
						"description": "Server error - Cognito or database error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Internal server error message",
											"example": "Internal server error"
										}
									},
									"required": ["message"]
								}
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"tags": ["Authentication"],
				"summary": "User login",
				"description": "Authenticates a user with email and password, returning access tokens upon successful login.",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email",
										"description": "User email address"
									},
									"password": {
										"type": "string",
										"minLength": 8,
										"maxLength": 15,
										"pattern": "\\d",
										"description": "User password"
									}
								},
								"required": ["email", "password"]
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "User logged in successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Response message"
										},
										"tokens": {
											"type": "object",
											"properties": {
												"accessToken": {
													"type": "string",
													"description": "JWT access token"
												},
												"refreshToken": {
													"type": "string",
													"description": "JWT refresh token"
												},
												"expiresIn": {
													"type": "number",
													"description": "Token expiration time in seconds"
												}
											},
											"required": ["accessToken", "refreshToken", "expiresIn"],
											"description": "Authentication tokens"
										}
									},
									"required": ["message", "tokens"],
									"description": "Authentication response"
								}
							}
						}
					},
					"400": {
						"description": "Invalid credentials or request data",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Validation error message",
											"example": "Validation Failed"
										},
										"details": {
											"type": "object",
											"additionalProperties": {
												"nullable": true
											},
											"description": "Detailed validation error information",
											"example": {
												"field": "email",
												"message": "Invalid email format"
											}
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized - Invalid email or password",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Unauthorized error message",
											"example": "Authentication required"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"403": {
						"description": "Forbidden - User not confirmed (UserNotConfirmedException)",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "User not confirmed error message",
											"example": "User is not confirmed"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"429": {
						"description": "Too many login attempts - rate limit exceeded",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"status": {
											"type": "number",
											"description": "HTTP status code",
											"example": 429
										},
										"message": {
											"type": "string",
											"description": "Rate limit error message",
											"example": "Too many requests, please try again later."
										}
									},
									"required": ["status", "message"]
								}
							}
						}
					},
					"500": {
						"description": "Server error - Cognito or database error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Internal server error message",
											"example": "Internal server error"
										}
									},
									"required": ["message"]
								}
							}
						}
					}
				}
			}
		},
...
		"/auth/logout": {
			"post": {
				"tags": ["Authentication"],
				"summary": "User logout",
				"description": "Logs out the authenticated user and invalidates their session.",
				"security": [
					{
						"cookieAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "User logged out successfully",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Response message"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized - Authentication required",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Unauthorized error message",
											"example": "Authentication required"
										}
									},
									"required": ["message"]
								}
							}
						}
					},
					"500": {
						"description": "Server error - Cognito or database error",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"description": "Internal server error message",
											"example": "Internal server error"
										}
									},
									"required": ["message"]
								}
							}
						}
					}
				}
			}
		},
...
	}
}